Updated page header styling Moved the customer header styling to styling folder so it can be used on multiple page components since it will be reused server times
