--- app/components/partNumbers/PartNumbers.css
@@ -1,61 +1,0 @@
-.container {
-  padding-top: 40px;
-  width: 100%;
-}
-
-.customer-head-container {
-  display: inline-flex;
-  width: 100%;
-  background-color: #8080802b;
-  padding: 10px 0;
-  position: fixed;
-}
-
-.customer-data {
-  margin-top: 100px;
-  margin-left: 15px;
-  overflow: hidden;
-}
-
-.container button {
-  font-family: serif;
-  border-radius: 3px;
-  filter: drop-shadow(1.5px 1.5px 1.5px black);
-  width: 140px;
-  margin: 5px;
-}
-
-.btnContainer button:hover {
-  background-color: cornflowerblue;
-  cursor: pointer;
-  transform: scale(1.01);
-}
-
-.btnContainer button:active {
-  background-color: cornflowerblue;
-  cursor: pointer;
-  box-shadow: 1px 1px rgba(255, 255, 255, 0.2);
-  transform: scale(0.97);
-  outline: 1px black;
-}
-
-.btnContainer button:focus {
-  outline: thin;
-}
-
-.btnContainer form {
-  margin: 5px;
-}
-
-.btnContainer form .form-container {
-  width: 100%;
-}
-
-.btnContainer form .form-container label {
-  font-size: 15px;
-}
-
-.btnContainer form .form-container input {
-  width: 100%;
-  padding: 2px;
-}
